Add files via upload
In this notebook, i began to explore the Events table in GDELT. I also was beginning to use some python libraries to try to do some basic data aggregation. At the time of creating this notebook, i was not very good in using the numpy and pandas python libraries. I also began to explore how to use SQL within python because i was already pretty fluent in SQL. Some of the notable things about this notebook was me attempting to graph my results at the end of the notebook and showing a time series graph of the number of mentions where China was actor 1 and the location was Guam for each month since 2015.