Add files via upload
In this notebook, I start trying to do some general tabulations of the south pacific like, how many events are there in the whole region? I also graph them using time-series to try to see if there is any strange anomalies. I also made a query where china is actor1 and Guam is the action location and graphed all the number of mentions from 2015 to 2020 for each month.